Self-Audited claim, fixed critical bug. Changed Variable names
Secured the minting function by changing the minter to account instead of msg.sender, where any one could have claimed for another person, as long as they had the correct proof and index and amount and account, now they are required to be the same person as the account they are claiming as (added a require statement and assigned the struct to the account parameter instead of msg.sender). Will be testing further!